extracted config_source from installer to be re-used in source command
