+++ conans/client/installer.py
@@ -4,7 +4,7 @@ import platform
 import fnmatch
 import shutil
 
-from conans.paths import CONANINFO, BUILD_INFO, DIRTY_FILE
+from conans.paths import CONANINFO, BUILD_INFO
 from conans.util.files import save, rmdir
 from conans.model.ref import PackageReference
 from conans.util.log import logger
@@ -14,8 +14,8 @@ from conans.client.generators import write_generators, TXTGenerator
 from conans.model.build_info import CppInfo
 from conans.client.output import ScopedOutput
 from conans.model.env_info import EnvInfo
-import six
 from conans.client.file_copier import report_copied_files
+from conans.client.source import config_source
 
 
 def init_info_objects(deps_graph, paths):
@@ -234,50 +234,6 @@ Package configuration:
         else:
             save(system_reqs_package_path, output)
 
-    def _config_source(self, export_folder, src_folder, conan_file, output):
-        """ creates src folder and retrieve, calling source() from conanfile
-        the necessary source code
-        """
-        dirty = os.path.join(src_folder, DIRTY_FILE)
-
-        def remove_source(raise_error=False):
-            output.warn("This can take a while for big packages")
-            try:
-                rmdir(src_folder, conan_file.short_paths)
-            except BaseException as e_rm:
-                save(dirty, "")  # Creation of DIRTY flag
-                msg = str(e_rm)
-                if six.PY2:
-                    msg = str(e_rm).decode("latin1")  # Windows prints some chars in latin1
-                output.error("Unable to remove source folder %s\n%s" % (src_folder, msg))
-                output.warn("**** Please delete it manually ****")
-                if raise_error or isinstance(e_rm, KeyboardInterrupt):
-                    raise ConanException("Unable to remove source folder")
-
-        if os.path.exists(dirty):
-            output.warn("Trying to remove dirty source folder")
-            remove_source(raise_error=True)
-        elif conan_file.build_policy_always:
-            output.warn("Detected build_policy 'always', trying to remove source folder")
-            remove_source(raise_error=True)
-
-        if not os.path.exists(src_folder):
-            output.info('Configuring sources in %s' % src_folder)
-            shutil.copytree(export_folder, src_folder)
-            save(dirty, "")  # Creation of DIRTY flag
-            os.chdir(src_folder)
-            try:
-                conan_file.source()
-                os.remove(dirty)  # Everything went well, remove DIRTY flag
-            except Exception as e:
-                os.chdir(export_folder)
-                # in case source() fails (user error, typically), remove the src_folder
-                # and raise to interrupt any other processes (build, package)
-                output.warn("Trying to remove dirty source folder")
-                remove_source()
-                msg = format_conanfile_exception(output.scope, "source", e)
-                raise ConanException(msg)
-
     def _build_package(self, export_folder, src_folder, build_folder, package_folder, conan_file,
                        output):
         """ builds the package, creating the corresponding build folder if necessary
@@ -287,7 +243,7 @@ Package configuration:
         """
         output.info('Building your package in %s' % build_folder)
         if not os.path.exists(build_folder):
-            self._config_source(export_folder, src_folder, conan_file, output)
+            config_source(export_folder, src_folder, conan_file, output)
             output.info('Copying sources to build folder')
 
             def check_max_path_len(src, files):
